+++ fastdsp.h
@@ -39,7 +39,6 @@
 #include "spectrum_window.h"
 
 #include "statistic_statsignal_F.h"
-#include "interpolation_envelope.h"
 #include "statistic_regonfly_F.h"