+++ filter_iir_F.h
@@ -17,9 +17,9 @@ namespace filter {
 typedef struct {
     float32         *weights_a;     //!< Recursive coefficients
     float32         *weights_b;     //!< FIR coefficients
-    float32         *buffer;        //!< Circular buffer
+    float32         *cbuffer;       //!< Circular buffer
     count_t         cb_index;       //!< Circular buffer index
-    count_t         size;       	//!< Circular buffer size. Size = max(size_a, size_b)
+    count_t         cb_size;       	//!< Circular buffer size. Size = max(size_a, size_b)
     count_t         size_a;         //!< Recursive weights count
     count_t         size_b;         //!< FIR weghts count. Count same as circular buffer elements count
     float32         input;          //!< Input value
@@ -33,6 +33,7 @@ EXTERN int iir_init_F(
         float32 *weights_a,
         float32 *weights_b,
         float32 *buffer,
+        count_t size,
         count_t size_a,
         count_t size_b
 );